se cambian el largo de algunas claves para guardarlas como hash, la creación de espectaculo se separa de las realizaciones de espectaculo

+++ src/main/java/com/grupo6/persistence/repository/TipoEspectaculoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.grupo6.persistence.model.TipoEspectaculo;
 
 @Repository
-public interface TipoEspectaculoRepository extends BaseRepository <TipoEspectaculo, Integer>{
+public interface TipoEspectaculoRepository extends BaseRepository <TipoEspectaculo, Long>{
 
 }
